+++ src/app/shared/directives/role.directive.ts
@@ -25,7 +25,7 @@ export class RoleDirective implements OnInit {
 
   @Input()
   set appRole(val: Array<string>) { // ['write']
-    // console.log(` ****`, val);
+    console.log(` ****`, val);
     this.permissions = val;
     this.updateView();
   }